Added harvey_nichols_auth_transaction_matching (AMEX) happy path test case
Added Scheme as a parameter for test case in iceland and HN, this will help make changes to merchant files where scheme is a mandatory parameter for file validation. changes are made to step definitions to capture "scheme"

Changed approval code in get_amex_auth_register_data to the required Amex standard removing the first 2 digits of the Auth code.